docker/pulsar/requirements.txt
pulsar-client
beautifulsoup4
googlesearch-python
nltk